pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aiodns>=4.0.0",
    "aiofiles>=25.1.0",
    "aiohttp>=3.13.3",
    "anyio>=4.12.1",
    "beautifulsoup4>=4.14.3",
    "brotli>=1.2.0",
    "cffconvert>=2.0.0",
    "fonttools>=4.61.1",
    "git-filter-repo>=2.47.0",
    "lxml>=6.0.2",
    "netaddr>=1.3.0",
    "pykwalify>=1.8.0",
    "pynacl>=1.6.2",
    "retrying>=1.4.2",
    "ruamel-yaml>=0.19.1",
    "shodan>=1.31.0",
    "theharvester>=0.0.1",
    "ujson>=5.12.0",
    "weasyprint>=68.1",
]
